Refactor example and existing sdk
