+++ PrivacySandboxKotlin/client-app/build.gradle
@@ -54,7 +54,7 @@ android {
 
 dependencies {
     // Depend on the SDK library that will run inside the app.
-    implementation project(':existing-sdk')
+    implementation project(':runtime-aware-sdk')
 
     implementation "androidx.appcompat:appcompat:1.6.1"
     implementation "com.google.android.material:material:1.9.0"
