+++ PrivacySandboxKotlin/client-app/src/main/java/com/example/client/MainActivity.kt
@@ -24,16 +24,16 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import com.example.privacysandbox.client.R
-import com.existing.sdk.BannerAd
-import com.existing.sdk.ExistingSdk
-import com.existing.sdk.FullscreenAd
+import com.runtimeaware.sdk.BannerAd
+import com.runtimeaware.sdk.ExistingSdk
+import com.runtimeaware.sdk.FullscreenAd
 import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
     /** Container for rendering content from the SDK. */
     private lateinit var bannerAd: BannerAd
 
-    private val existingSdk = ExistingSdk(this)
+    private val runtimeAwareSdk = ExistingSdk(this)
 
     /** A spinner for selecting the size of the file created in the sandbox. */
     private lateinit var fileSizeSpinner: Spinner
@@ -118,7 +118,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun onInitializeSkButtonPressed() = lifecycleScope.launch {
-        if (!existingSdk.initialize()) {
+        if (!runtimeAwareSdk.initialize()) {
             makeToast("Failed to initialize SDK")
         } else {
             makeToast("Initialized SDK!")
@@ -158,7 +158,7 @@ class MainActivity : AppCompatActivity() {
         val fileSize = fileSizes[fileSizeSpinner.selectedItemPosition]
 
         lifecycleScope.launch {
-            val success = existingSdk.createFile(fileSize.sizeInMb)
+            val success = runtimeAwareSdk.createFile(fileSize.sizeInMb)
             if (success == null) {
                 makeToast("Please load the SDK first!")
                 return@launch
