--- PrivacySandboxKotlin/example-sdk/build.gradle
@@ -1,71 +1,0 @@
-/*
- * Copyright (C) 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-plugins {
-    id "androidx.privacysandbox.library"
-    id "org.jetbrains.kotlin.android"
-}
-
-android {
-    namespace "com.example"
-
-    defaultConfig {
-        minSdk 21
-        compileSdk 34
-        compileSdkExtension 12
-        targetSdk 34
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro"
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_17
-        targetCompatibility JavaVersion.VERSION_17
-    }
-    kotlinOptions {
-        jvmTarget = '17'
-    }
-}
-
-dependencies {
-    implementation 'androidx.activity:activity-ktx:1.9.0'
-    ksp 'androidx.annotation:annotation:1.6.0'
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.10'
-    implementation "androidx.lifecycle:lifecycle-common:2.7.0"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.1"
-
-    implementation "androidx.privacysandbox.activity:activity-client:$privacy_sandbox_activity_version"
-    implementation "androidx.privacysandbox.activity:activity-core:$privacy_sandbox_activity_version"
-    implementation "androidx.privacysandbox.activity:activity-provider:$privacy_sandbox_activity_version"
-
-    implementation "androidx.privacysandbox.ui:ui-core:$privacy_sandbox_ui_version"
-    implementation "androidx.privacysandbox.ui:ui-provider:$privacy_sandbox_ui_version"
-    implementation "androidx.privacysandbox.ui:ui-client:$privacy_sandbox_ui_version"
-
-    implementation "androidx.privacysandbox.sdkruntime:sdkruntime-core:$privacy_sandbox_sdk_runtime_version"
-    implementation "androidx.privacysandbox.sdkruntime:sdkruntime-client:$privacy_sandbox_sdk_runtime_version"
-    implementation "androidx.privacysandbox.sdkruntime:sdkruntime-provider:$privacy_sandbox_sdk_runtime_version"
-
-    ksp "androidx.privacysandbox.tools:tools-apicompiler:$privacy_sandbox_tools_version"
-    implementation "androidx.privacysandbox.tools:tools:$privacy_sandbox_tools_version"
-}
