--- PrivacySandboxKotlin/example-sdk/src/main/java/com/example/api/SdkBannerRequest.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.api
-
-import androidx.privacysandbox.tools.PrivacySandboxValue
-import androidx.privacysandbox.activity.core.SdkActivityLauncher
-
-@PrivacySandboxValue
-data class SdkBannerRequest(
-    /** The package name of the app. */
-    val appPackageName: String,
-    /**
-     *  An [SdkActivityLauncher] that will be used to launch an activity when the banner is clicked.
-     */
-    val activityLauncher: SdkActivityLauncher,
-    /**
-     * Denotes if a WebView banner ad needs to be loaded.
-     */
-    val isWebViewBannerAd: Boolean
-)
