+++ PrivacySandboxKotlin/inapp-mediatee-sdk-adapter/README.md
@@ -3,7 +3,7 @@
 This SDK is Runtime Aware but runs in the App process. It facilitates interaction between mediator
 and in-app mediatee.
 
-Implements MediateeAdapterInterface declared by mediator (example-sdk).
+Implements MediateeAdapterInterface declared by mediator (runtime-enabled).
 
 This could be owned by the mediator sdk during transition, or optionally all the logic here could
-also be a part of RA_SDK (existing-sdk).
+also be a part of RA_SDK (runtime-aware-sdk).
