+++ PrivacySandboxKotlin/inapp-mediatee-sdk-adapter/build.gradle
@@ -46,7 +46,7 @@ android {
 }
 
 dependencies {
-    debugImplementation project(':example-sdk-bundle')
+    debugImplementation project(':runtime-enabled-sdk-bundle')
     implementation project(':inapp-mediatee-sdk')
 
     implementation 'androidx.activity:activity-ktx:1.9.0'
