+++ PrivacySandboxKotlin/runtime-aware-sdk/src/main/AndroidManifest.xml
