+++ PrivacySandboxKotlin/runtime-enabled-sdk/src/main/java/com/runtimeenabled/api/FullscreenAd.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.runtimeenabled.api
+
+import androidx.privacysandbox.activity.core.SdkActivityLauncher
+import androidx.privacysandbox.tools.PrivacySandboxInterface
+
+@PrivacySandboxInterface
+interface FullscreenAd {
+    suspend fun show(activityLauncher: SdkActivityLauncher)
+}
