+++ PrivacySandboxKotlin/runtime-enabled-sdk/src/main/res/layout/full_screen.xml
