+++ PrivacySandboxKotlin/settings.gradle
@@ -33,12 +33,12 @@ dependencyResolutionManagement {
 
 rootProject.name = "PrivacySandboxSample"
 include ':client-app'
-include ':example-sdk'
-include ':example-sdk-bundle'
+include ':runtime-enabled-sdk'
+include ':runtime-enabled-sdk-bundle'
 include ':inapp-mediatee-sdk'
 include ':inapp-mediatee-sdk-adapter'
 include ':mediatee-sdk'
 include ':mediatee-sdk-bundle'
 include ':mediatee-sdk-adapter'
 include ':mediatee-sdk-adapter-bundle'
-include ':existing-sdk'
+include ':runtime-aware-sdk'
